requirements.txt
absl-py==1.0.0
albumentations==1.2.1
backcall==0.2.0
blis==0.7.5
cachetools==4.2.4
catalogue==2.0.6
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==1.6.0
configy==0.1.4
contextvars==2.4
cycler==0.10.0
cymem==2.0.6
Cython==0.29.23
dataclasses==0.8
decorator==4.4.2
dominate==2.6.0
entrypoints==0.4
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grad-cam==1.4.5
grpcio==1.42.0
idna==2.6
imageio==2.9.0
immutables==0.16
importlib-metadata==4.8.2
install==1.3.5
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.1
langcodes==3.3.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
murmurhash==1.0.6
nest-asyncio==1.5.5
netron==5.9.9
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.1.5
parso==0.7.1
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
preshed==3.0.6
prompt-toolkit==3.0.30
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pydantic==1.8.2
Pygments==2.13.0
PyGObject==3.26.1
pyparsing==2.4.7
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.1
python-distutils-extra==2.39
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==23.2.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
ssh-import-id==5.7
thinc==8.0.13
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.5.4
torch==1.10.1
torch-tb-profiler==0.3.1
torchfile==0.1.0
torchvision==0.11.2
tornado==6.1
tqdm==4.61.1
traitlets==4.3.3
ttach==0.0.3
typer==0.4.0
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.22
visdom==0.1.8.8
wasabi==0.9.0
wcwidth==0.2.5
Werkzeug==2.0.2
yacs==0.1.8
zipp==3.6.0
